requirements.txt
absl-py==2.1.0
asttokens @ file:///home/conda/feedstock_root/build_artifacts/asttokens_1733250440834/work
black @ file:///home/conda/feedstock_root/build_artifacts/black-recipe_1728503730172/work
build==1.2.2.post1
click @ file:///home/conda/feedstock_root/build_artifacts/click_1733221831880/work
ConfigArgParse==1.7
contourpy==1.3.1
cycler==0.12.1
decorator @ file:///home/conda/feedstock_root/build_artifacts/decorator_1733236420667/work
exceptiongroup @ file:///home/conda/feedstock_root/build_artifacts/exceptiongroup_1733208806608/work
executing @ file:///home/conda/feedstock_root/build_artifacts/executing_1733569351617/work
filelock==3.16.1
fonttools==4.55.2
fsspec==2024.10.0
grpcio==1.68.1
imageio==2.36.1
imageio-ffmpeg==0.5.1
ipdb @ file:///home/conda/feedstock_root/build_artifacts/ipdb_1679608408213/work
ipython @ file:///home/conda/feedstock_root/build_artifacts/ipython_1732896932739/work
jedi @ file:///home/conda/feedstock_root/build_artifacts/jedi_1733300866624/work
Jinja2==3.1.4
kiwisolver==1.4.7
Markdown==3.7
MarkupSafe==3.0.2
matplotlib==3.9.3
matplotlib-inline @ file:///home/conda/feedstock_root/build_artifacts/matplotlib-inline_1733416936468/work
mpmath==1.3.0
mypy_extensions @ file:///home/conda/feedstock_root/build_artifacts/mypy_extensions_1733230897951/work
networkx==3.4.2
numpy==2.1.3
nvidia-cublas-cu12==12.4.5.8
nvidia-cuda-cupti-cu12==12.4.127
nvidia-cuda-nvrtc-cu12==12.4.127
nvidia-cuda-runtime-cu12==12.4.127
nvidia-cudnn-cu12==9.1.0.70
nvidia-cufft-cu12==11.2.1.3
nvidia-curand-cu12==10.3.5.147
nvidia-cusolver-cu12==11.6.1.9
nvidia-cusparse-cu12==12.3.1.170
nvidia-nccl-cu12==2.21.5
nvidia-nvjitlink-cu12==12.4.127
nvidia-nvtx-cu12==12.4.127
opencv-python==4.10.0.84
packaging @ file:///home/conda/feedstock_root/build_artifacts/packaging_1733203243479/work
parso @ file:///home/conda/feedstock_root/build_artifacts/parso_1733271261340/work
pathspec @ file:///home/conda/feedstock_root/build_artifacts/pathspec_1733233363808/work
pexpect @ file:///home/conda/feedstock_root/build_artifacts/pexpect_1733301927746/work
pickleshare @ file:///home/conda/feedstock_root/build_artifacts/pickleshare_1733327343728/work
pillow==11.0.0
pip-tools==7.4.1
platformdirs @ file:///home/conda/feedstock_root/build_artifacts/platformdirs_1733232627818/work
prompt_toolkit @ file:///home/conda/feedstock_root/build_artifacts/prompt-toolkit_1733302527033/work
protobuf==5.29.1
ptyprocess @ file:///home/conda/feedstock_root/build_artifacts/ptyprocess_1733302279685/work/dist/ptyprocess-0.7.0-py2.py3-none-any.whl#sha256=92c32ff62b5fd8cf325bec5ab90d7be3d2a8ca8c8a3813ff487a8d2002630d1f
pure_eval @ file:///home/conda/feedstock_root/build_artifacts/pure_eval_1733569405015/work
pyglet==1.5.30
Pygments @ file:///home/conda/feedstock_root/build_artifacts/pygments_1733221634316/work
PyMCubes==0.1.6
pyparsing==3.2.0
pyproject_hooks==1.2.0
python-dateutil==2.9.0.post0
scipy==1.14.1
setuptools==75.6.0
six==1.17.0
stack_data @ file:///home/conda/feedstock_root/build_artifacts/stack_data_1733569443808/work
sympy==1.13.1
tensorboard==2.18.0
tensorboard-data-server==0.7.2
toml @ file:///home/conda/feedstock_root/build_artifacts/toml_1604308577558/work
torch==2.5.1
torchvision==0.20.1
tqdm==4.67.1
traitlets @ file:///home/conda/feedstock_root/build_artifacts/traitlets_1733367359838/work
trimesh==4.5.3
triton==3.1.0
typing_extensions @ file:///home/conda/feedstock_root/build_artifacts/typing_extensions_1733188668063/work
wcwidth @ file:///home/conda/feedstock_root/build_artifacts/wcwidth_1733231326287/work
Werkzeug==3.1.3
wheel==0.45.1
